FP-0009 fixing bugs noticed by leader
