+++ src/app/app.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
-//import { StoreModule } from '@ngrx/store';
+import { StoreModule } from '@ngrx/store';
 import { AppComponent } from './app.component';
 import { MainComponent } from './main/main.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { SharedModule } from './shared/shared.module';
 import { UsersModule } from './users/users.module';
-import { MatToolbarModule } from '@angular/material/toolbar';
-//import * as fromApp from './store/app.reducer';
+import * as fromApp from './store/app.reducer';
 
 @NgModule({
   declarations: [AppComponent, MainComponent],
@@ -17,9 +17,9 @@ import { MatToolbarModule } from '@angular/material/toolbar';
     AppRoutingModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    //StoreModule.forRoot(fromApp.appReducer),
+    StoreModule.forRoot(fromApp.appReducer),
+    SharedModule,
     UsersModule,
-    MatToolbarModule,
   ],
   providers: [],
   bootstrap: [AppComponent],
