+++ src/app/users/users-delete-confirm.html
@@ -1,4 +1,4 @@
-<h2 mat-dialog-title>Install Angular</h2>
+<h2 mat-dialog-title>Confirm Notice</h2>
 <mat-dialog-content class="mat-typography">
   <p>Are you sure you want to delete?</p>
 </mat-dialog-content>
