+++ src/app/users/users.component.html
@@ -20,19 +20,14 @@
   </ng-container>
 
   <!-- edit Column -->
-  <ng-container matColumnDef="edit">
-    <th mat-header-cell *matHeaderCellDef>Edit</th>
+  <ng-container matColumnDef="actions">
+    <th mat-header-cell *matHeaderCellDef>Actions</th>
     <td mat-cell *matCellDef="let element" [@EnterLeave]>
       <button mat-mini-fab color="primary" (click)="onEdit(element.id)">
         Edit
       </button>
-    </td>
-  </ng-container>
-  <!-- Delete Column -->
-  <ng-container matColumnDef="delete">
-    <th mat-header-cell *matHeaderCellDef>Delete</th>
-    <td mat-cell *matCellDef="let element" [@EnterLeave]>
-      <button mat-mini-fab color="warn" (click)="openDialog(element.id)">
+      |
+      <button mat-mini-fab color="warn" (click)="openDeleteDialog(element.id)">
         Delete
       </button>
     </td>
